Criação das camadas resource/controller, service e mapper.

+++ backend/app/src/main/java/com/appproduto/repository/ProdutoRepository.java
@@ -3,5 +3,5 @@ package com.appproduto.repository;
 import com.appproduto.model.ProdutoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<ProdutoModel, Long> {
+public interface ProdutoRepository extends JpaRepository<ProdutoModel, Integer> {
 }
